Pass in `io.Reader` instead of `[]byte` to `NewScanner`
io.Reader is a more generic abstraction for input: it could be a file or it could be a reader of stdin.

The changes will enable `ccat` to use the `NewScannaer` API to create a `Scanner`. See https://github.com/jingweno/ccat/commit/e7385e67afd8690c1ea677abfc13f2807aadfe05.